Option to specify 'sqwt' or 'rixs' data for analysis module.
The analysis module is now given the option where the user can
specify whether the data is a structure factor calculation ('sqwt')
or it is a trRIXS ('rixs') calculation. Once it is told, the QFI
spit out by the routine has all the appropirate prefactors taken
care of. The user will only have to divide by the number of the
sites to get the QFI density.